Migrate security test into proper test file after 2.252
