+++ test/src/test/resources/hudson/model/CauseTest/jobs/fs/config.xml
